+++ clumio/plugin_framework/clumio_protection_group_asset/const.go
@@ -1,0 +1,12 @@
+// Copyright 2024. Clumio, Inc.
+
+package clumio_protection_group_asset
+
+const (
+	// Constants used by the resource model for the clumio_protection_group_asset Terraform
+	// resource. These values should match the schema tfsdk tags on the resource model struct in
+	// schema.go.
+	schemaId                = "id"
+	schemaBucketId          = "bucket_id"
+	schemaProtectionGroupId = "protection_group_id"
+)