+++ examples/data-sources/clumio_role/data-source.tf
@@ -1,2 +1,3 @@
 data "clumio_role" "example" {
+  name = "Super Admin"
 }